friendly fire doesnt provide xp career ability usage doesnt provide any xp
moved AttackTypeMask to DamageHelper

removed unnecessary comments
